Adds SolrDocs and SlowSolrDocs in sourcSource module
As per @joekiller's feedback, its useful to keep both strategies around.
The slower version is frought with deep paging, but its more likely to
work on any Solr.
